refactor: wkc layout — phase 13 (drain non-cache ports)
migrates the three legacy ports/ files that don't depend on an
in-memory cache out of src/ports/ and into proper wkc adapter folders:

- ports/postgres.ts → adapters/database/ (component.ts, types.ts,
  index.ts). database client + transactional client + idatabasecomponent
  interface live in types.ts. createtestdatabasecomponent moves to
  test/mocks/database-component-mock.ts so the production adapter no
  longer ships a test stub.
- ports/snapshotstorage.ts → adapters/snapshot-storage/. thin facade
  over snapshots-repository.isownsnapshot. types.ts re-exports the
  external isnapshotstoragecomponent interface as snapshotstorage.
- ports/snapshotgenerator.ts → adapters/snapshot-generator/. preserves
  the existing scheduled-job lifecycle (start/stop) and the
  getcurrentsnapshots accessor.

repoints all import sites (22 in src/, ~9 in test/) and relocates the
postgres / snapshotstorage / snapshotgenerator specs under test to
mirror the new layout. cleans up the components.ts header now that
three of the five legacy entries are gone.

failed-deployments and processed-snapshot-storage are deferred to a
follow-up that lands together with the @dcl/memory-cache-component
minor bump (core-components pr #88).

233 unit tests passing.
