+++ content/src/adapters/database/index.ts
@@ -1,0 +1,2 @@
+export { createDatabase, createDatabaseComponent } from './component'
+export type { DatabaseClient, DatabaseTransactionalClient, IDatabaseComponent } from './types'
