+++ content/src/adapters/pointers-repository/types.ts
@@ -1,4 +1,4 @@
-import { DatabaseClient } from '../../ports/postgres'
+import { DatabaseClient } from '../../adapters/database'
 
 export interface IPointersRepository {
   getItemEntitiesIdsThatMatchCollectionUrnPrefix(db: DatabaseClient, collectionUrn: string): Promise<string[]>
