+++ content/src/adapters/snapshot-generator/index.ts
@@ -1,0 +1,2 @@
+export { createSnapshotGenerator } from './component'
+export type { SnapshotGenerator } from './types'
