+++ content/src/adapters/snapshot-generator/types.ts
@@ -1,0 +1,6 @@
+import { SnapshotMetadata } from '@dcl/snapshots-fetcher/dist/types'
+import { IBaseComponent } from '@well-known-components/interfaces'
+
+export type SnapshotGenerator = IBaseComponent & {
+  getCurrentSnapshots(): SnapshotMetadata[] | undefined
+}
