+++ content/src/adapters/snapshot-storage/index.ts
@@ -1,0 +1,2 @@
+export { createSnapshotStorage } from './component'
+export type { SnapshotStorage } from './types'
