+++ content/src/adapters/snapshot-storage/types.ts
@@ -1,0 +1,3 @@
+import { ISnapshotStorageComponent } from '@dcl/snapshots-fetcher/dist/types'
+
+export type SnapshotStorage = ISnapshotStorageComponent
