--- content/src/ports/snapshotStorage.ts
@@ -1,10 +1,0 @@
-import { isOwnSnapshot } from '../adapters/snapshots-repository'
-import { AppComponents } from '../types'
-
-export function createSnapshotStorage(components: Pick<AppComponents, 'database'>) {
-  return {
-    async has(snapshotHash: string) {
-      return isOwnSnapshot(components.database, snapshotHash)
-    }
-  }
-}
