+++ content/src/types.ts
@@ -34,9 +34,9 @@ import { DeployedEntitiesBloomFilter } from './adapters/deployed-entities-bloom-
 import { Deployer } from './logic/deployment-service'
 import { IPointerLockManager } from './adapters/pointer-lock-manager'
 import { IFailedDeploymentsComponent } from './ports/failedDeployments'
-import { IDatabaseComponent } from './ports/postgres'
+import { IDatabaseComponent } from './adapters/database'
 import { ISequentialTaskExecutorComponent } from './logic/sequential-task-executor'
-import { SnapshotGenerator } from './ports/snapshotGenerator'
+import { SnapshotGenerator } from './adapters/snapshot-generator'
 import { SynchronizationState } from './adapters/synchronization-state'
 import { SystemProperties } from './adapters/system-properties'
 import { GarbageCollectionManager } from './logic/garbage-collection'
