+++ content/test/integration/E2ETestEnvironment.ts
@@ -10,7 +10,7 @@ import { stopAllComponents } from '../../src/logic/components-lifecycle'
 import { metricsDeclaration } from '../../src/metrics'
 import { createMigrationExecutor } from '../../src/migrations/migration-executor'
 import { DAOComponent } from '../../src/adapters/dao-client'
-import { IDatabaseComponent, createDatabaseComponent } from '../../src/ports/postgres'
+import { IDatabaseComponent, createDatabaseComponent } from '../../src/adapters/database'
 import { AppComponents } from '../../src/types'
 import { MockedDAOClient } from '../helpers/service/synchronization/clients/MockedDAOClient'
 import { createNoOpDeployRateLimiter } from '../mocks/deploy-rate-limiter-mock'
