+++ content/test/integration/adapters/snapshot-generator.spec.ts
