+++ content/test/integration/adapters/snapshot-storage.spec.ts
