+++ content/test/mocks/database-component-mock.ts
@@ -1,4 +1,4 @@
-import { IDatabaseComponent } from '../../src/ports/postgres'
+import { IDatabaseComponent } from '../../src/adapters/database'
 
 export const createDatabaseMockedComponent = (
   overrides?: Partial<jest.Mocked<IDatabaseComponent>>
@@ -11,5 +11,24 @@ export const createDatabaseMockedComponent = (
     start: jest.fn(),
     stop: jest.fn(),
     ...overrides
+  }
+}
+
+export function createTestDatabaseComponent(): IDatabaseComponent {
+  return {
+    async query() {
+      throw new Error('query Not implemented')
+    },
+    async queryWithValues() {
+      throw new Error('queryWithValues Not implemented')
+    },
+    async *streamQuery() {
+      throw new Error('streamQuery Not implemented')
+    },
+    async transaction() {
+      throw new Error('transactionQuery Not implemented')
+    },
+    async start() {},
+    async stop() {}
   }
 }
