+++ content/test/unit/ports/failedDeployments.spec.ts
@@ -3,7 +3,8 @@ import { createTestMetricsComponent } from '@dcl/metrics'
 import * as failedDeploymentQueries from '../../../src/adapters/failed-deployments-repository'
 import { metricsDeclaration } from '../../../src/metrics'
 import { createFailedDeployments, FailureReason, SnapshotFailedDeployment } from '../../../src/ports/failedDeployments'
-import { createTestDatabaseComponent, DatabaseTransactionalClient } from '../../../src/ports/postgres'
+import { createTestDatabaseComponent } from '../../mocks/database-component-mock'
+import { DatabaseTransactionalClient } from '../../../src/adapters/database'
 import { AppComponents } from '../../../src/types'
 
 describe('failed deployments', () => {
